+++ OndselSolver/ASMTSpatialItem.h
@@ -31,8 +31,8 @@ namespace MbD {
 		void storeOnLevel(std::ofstream& os, int level) override;
 		void storeOnLevelPosition(std::ofstream& os, int level);
 		void storeOnLevelRotationMatrix(std::ofstream& os, int level);
-		FColDsptr getPosition3D(size_t i);
-		FMatDsptr getRotationMatrix(size_t i);
+		FColDsptr getPosition3D(int i);
+		FMatDsptr getRotationMatrix(int i);
 
 		FColDsptr position3D = std::make_shared<FullColumn<double>>(3);
 		FMatDsptr rotationMatrix = std::make_shared<FullMatrixDouble>(ListListD{
